LOAD/STORE instruction test passed
It's so important that the NOP instruction must be implemented.
Regarding NOP as a special case of SLL would break at some tests.
Because SLL set WriteEnable and it may happen that when RAW with a
NOP instruction,zero would be put to Decode Phase and cause serious
problems.

Another note is that when ZZZZZ signals are found in the wave,it is
highly caused by signal connection missing.